requirements.txt
pip~=21.2.4
setuptools~=58.1.0
celery==5.4.0
fastapi==0.111.1
joblib==1.4.2
pandas==2.2.2
pika==1.3.2
prophet==1.1.5
pydantic==2.8.
pyTelegramBotAPI==4.21.0
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.0
uvicorn==0.30.3
psycopg2-binary
alembic
numpy<2
python-jose
passlib
streamlit<1.37.0
matplotlib
streamlit_cookies_manager
python-dotenv
requests-mock
telemulator3
jose
pytest
catboost
aiogram
pandas
json
re
pypdf
shap
joblib